requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tweepy>=4.14.0
python-dateutil>=2.8.0
jinja2>=3.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiosqlite>=0.19.0
